% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMEA_get.R
\name{ccfResNames}
\alias{ccfResNames}
\title{Extract the names of the ccf analysis summaries in a MEA objects}
\usage{
ccfResNames(mea)
}
\arguments{
\item{mea}{an object of class \code{MEA} or a list of \code{MEA} objects (see function \code{\link{readMEA}})}
}
\value{
a vector containing the labels of the ccfRes indexes
}
\description{
Extract the names of the ccf analysis summaries in a MEA objects
}
