% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMEA_graphics.R
\name{colTrans}
\alias{colTrans}
\title{Transform color}
\usage{
colTrans(col, luminosity = NA, alpha = NA)
}
\arguments{
\item{col}{a color to begin with in hex format}

\item{luminosity}{numeric. negative numbers darken the color, positive lighten it. Eg: a value of -2 make the color two times darker.}

\item{alpha}{numeric from 0 to 1. the value of opacity of the resulting color}
}
\value{
a color string
}
\description{
Transform color
}
