% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMEA_get.R
\name{lagNames}
\alias{lagNames}
\title{Extract the lag names of a ccf analysis in MEA objects}
\usage{
lagNames(mea)
}
\arguments{
\item{mea}{an object of class \code{MEA} or a list of \code{MEA} objects (see function \code{\link{readMEA}})}
}
\value{
a vector containing the labels of the lag values
}
\description{
Extract the lag names of a ccf analysis in MEA objects
}
