% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMEA_class.R
\name{MEAlist}
\alias{MEAlist}
\alias{is.MEAlist}
\title{Well formatted list of \code{MEA} objects}
\usage{
MEAlist(listOfMea)

is.MEAlist(x)
}
\arguments{
\item{listOfMea}{a list containing \code{MEA} objects}

\item{x}{object to be tested.}
}
\value{
an object of class \code{MEAlist}

\code{is.MEAlist} returns TRUE if and only if its argument is of class \code{MEAlist}
}
\description{
This constructor function checks if all the supplied \code{MEA} objects share the same sampling rate, pre-processing, and metadata,
and returns an object with additional attributes summarizing the contained \code{MEA} objects.
}
