% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMEA_graphics.R
\name{plot.MEAlist}
\alias{plot.MEAlist}
\title{Plots all MEA signals contained in an object of class \code{MEAlist}}
\usage{
\method{plot}{MEAlist}(x, from = 1, to = max(sapply(mea, function(x) {    
  nrow(x$MEA) })), ccf = F, rescale = F, ...)
}
\arguments{
\item{x}{an object of class \code{MEAlist}}

\item{from}{integer. The first sample to be plotted. Defaults to the beginning of the signals.}

\item{to}{integer. The last sample to be plotted. Defaults to the length of the longest MEA signal}

\item{ccf}{either FALSE or a string representing the type of ccf to be overlayed.
One of "sync", "pace", "zero", "lead", "pace0", "lead0".}

\item{rescale}{logical. Should the motion energy time-series be rescaled?}

\item{...}{further graphical parameters passed to  \code{\link[graphics]{plot}}}
}
\description{
Plots all MEA signals contained in an object of class \code{MEAlist}
}
\details{
Note: if more of than 10s of trailing zeroes are found at the end of both s1 and s2 signals they are truncated.
}
