% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMEA_get.R
\name{id}
\alias{id}
\alias{group}
\alias{session}
\alias{sampRate}
\alias{s1Name}
\alias{s2Name}
\alias{uid}
\title{Get MEA attributes}
\usage{
id(mea)

group(mea)

session(mea)

sampRate(mea)

s1Name(mea)

s2Name(mea)

uid(mea)
}
\arguments{
\item{mea}{a single or a list of MEA objects}
}
\value{
A string or a vector of strings containing the metadata.
}
\description{
Get MEA attributes
}
\details{
if a well formatted list of MEA objects is provided, the function returns a vector of results.
}
